+++ src/main/java/org/multiagent/strategies/Strategie.java
@@ -4,5 +4,5 @@ import org.multiagent.communication.Negociation;
 
 public interface Strategie {
 
-    public boolean appliquer(Negociation negociation, Float objective_price);
+    public boolean appliquer(Negociation negociation, Double objective_price);
 }
